refactor(db): extract chat session, message, and attachment domain into db/chat.rs
